+++ gosdk-wrapper-maven-plugin/src/main/java/com/igormaznitsa/mvngolang/utils/ApacheHttpClient5Loader.java
@@ -265,7 +265,7 @@ public class ApacheHttpClient5Loader {
     }
 
     return builder
-        .setUserAgent("gosdk-wrapper-maven-plugin-agent/1.0.6")
+        .setUserAgent("gosdk-wrapper-maven-plugin-agent/1.1.2")
         .disableCookieManagement()
         .build();
   }